azure_func_api/requirements.txt
azure-functions
python-dotenv
lancedb
sentence-transformers
google-genai
